app/server/requirements-vercel.txt
# Vercel-compatible requirements (no heavy ML dependencies)
# easyocr, torch, opencv, scikit-image, etc. are excluded because
# they exceed Vercel's 250MB serverless function size limit.
#
# The OCR/verify-id endpoint will NOT work on Vercel.
# All data/analytics endpoints work fine.

fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
python-multipart==0.0.22
pandas==2.3.3
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
supabase==2.28.0
rapidfuzz==3.14.3
numpy==2.4.1
requests==2.32.5
